[frontend] Refactoring on the works. i18n strings are broken and angular is failing due to this.
