+++ ethicapp/frontend/assets/js/controllers/teacher/routing_controller.js
@@ -1,0 +1,15 @@
+/*eslint func-style: ["error", "expression"]*/
+export let RoutingController = ($scope) => {
+    $scope.template = {      
+        "home":           "views/partials/teacher/home.html",
+        "newDesign":      "views/partials/teacher/new-design.html",
+        "newDesignExt":   "views/partials/teacher/new-design-ext.html",
+        "designs":        "views/partials/teacher/designs.html",
+        "users":          "views/partials/teacher/users.html",
+        "activities":     "views/partials/teacher/activities.html",
+        "launchActivity": "views/partials/teacher/launch-activity.html",
+        "viewDesign":     "views/partials/teacher/view-design.html",
+        "activity":       "views/partials/teacher/activity.html",
+        "profile":        "views/partials/profile.html",
+    };
+};
