+++ ethicapp/frontend/views/deprecated/admin.ejs
@@ -11,7 +11,7 @@
     <script src="assets/libs/angular-nvd3.min.js"></script>
     <script type="text/javascript" src="assets/libs/ng-quill.min.js"></script>
     <script src="assets/js/StagesController.js"></script>
-    <script src="assets/js/admin-ctrl.js"></script>
+    <script src="assets/js/ngmodules/teacher/teacher_admin.js"></script>
     <script src="/socket.io/socket.io.js"></script>
     <script src="libs/ngDialog.min.js"></script>
     <script src="<%= asset('/assets/js/assets-bundle.js') %>"></script>
