+++ ethicapp/frontend/views/profile.ejs
@@ -10,7 +10,7 @@
     <script src="assets/libs/angular.min.js"></script>
     <script src="assets/libs/angular-nvd3.min.js"></script>
     <script type="text/javascript" src="assets/libs/ng-quill.min.js"></script>
-    <script src="assets/js/StagesController.js"></script>
+    <script src="assets/js/ngmodules/teacher_admin.js"></script>
     <script src="assets/js/admin-ctrl.js"></script>
     <script src="/socket.io/socket.io.js"></script>
     <script src="<%= asset('/js/assets-bundle.js') %>"></script>
